+++ ooni/utils/__init__.py
@@ -1,12 +1,10 @@
 import shutil
 import string
 import random
-import glob
 import os
 from datetime import datetime
 
 import gzip
-from base64 import b64encode
 from zipfile import ZipFile
 
 from ooni import errors
@@ -155,9 +153,3 @@ def gunzip(filename, dst):
 def get_ooni_root():
     script = os.path.join(__file__, '..')
     return os.path.dirname(os.path.realpath(script))
-
-def base64Dict(data):
-    return {
-        'format': 'base64',
-        'data': b64encode(data)
-    }
